# Frontend improvements and audio optimization