--- a/landing/.gitignore
+++ b/landing/.gitignore
@@ -32,3 +32,6 @@ yarn-error.log*
 # typescript
 *.tsbuildinfo
 next-env.d.ts
+
+# generated docs
+/app/docs/content/