--- a/src-tauri/src/audio/mod.rs
+++ b/src-tauri/src/audio/mod.rs
@@ -2,7 +2,12 @@ pub mod recorder;
 pub mod converter;
 pub mod ring_buffer_recorder;
 pub mod format;
+pub mod metadata;
+
+#[cfg(test)]
+mod test_metadata;
 
 pub use recorder::AudioRecorder;
 pub use converter::AudioConverter;
 pub use format::WhisperAudioConverter;
+pub use metadata::AudioMetadata;