… .claude/agents/backend-engineer-david.md
… .claude/agents/product-designer-jimmy.md
… .claude/settings.local.json
… DESIGN_IMPROVEMENTS.md
… docs/AUDIO_METADATA.md
… docs/AUDIO_PIPELINE.md
… docs/PIPELINE_OVERVIEW.md
… docs/README.md
… docs/audio-recorder-changes.md
… landing/app/docs/docs-page-content.tsx
… landing/app/docs/page.tsx
… landing/package.json
… src-tauri/src/audio/format.rs
… src-tauri/src/audio/metadata.rs
… src-tauri/src/audio/notifications.rs
… src-tauri/src/audio/recorder.rs
… src-tauri/src/audio/test_metadata.rs
… src-tauri/src/audio/validation.rs
… src-tauri/src/audio/wav_validator.rs
… src-tauri/src/db/mod.rs
… src-tauri/src/keyboard_monitor.rs
… src-tauri/src/lib.rs
… src-tauri/src/llm/engine.rs
… src-tauri/src/llm/models.rs
… src-tauri/src/macos/app_context.rs
… src-tauri/src/macos/window_ext.rs
… src-tauri/src/performance_metrics_service.rs
… src-tauri/src/post_processing.rs
… src-tauri/src/recording_workflow.rs
… src-tauri/src/ring_buffer_monitor.rs