## Summary

This PR includes significant frontend UI/UX improvements and audio recording optimizations that reduce file sizes by ~12x.

## Key Changes

### 🎨 Frontend UI/UX Improvements
- **Header/Sidebar Refactoring**: Moved sidebar toggle into sidebar itself for better visual hierarchy
- **Visual Polish**: Fixed transcript group header backgrounds, improved hover states
- **Light Theme**: Enhanced contrast and readability for light theme users
- **Documentation**: Added DESIGN_IMPROVEMENTS.md tracking all UI enhancements

### ⚡️ Audio Optimization
- **12x smaller files**: Implemented 16kHz mono recording (750KB vs 9MB for 25 seconds)
- **Direct capture**: Records at optimal format without post-processing conversion
- **Memory savings**: 66% reduction in ring buffer memory usage
- **Graceful fallback**: Supports devices that don't have 16kHz mono capability

### ✨ New Features
- **WAV Export**: Added ability to download audio files from transcript detail page
- **Audio Monitoring**: Foundation for device change detection and notifications

### 🐛 Bug Fixes
- Fixed Tiny model hallucinations by switching to Medium model with 10s chunks
- Fixed WAV export security restrictions with backend command approach
- Reduced audio cutoff threshold from 500ms to 100ms

## Testing
- Audio recording tested with various devices
- File size reduction verified
- UI improvements tested in both light and dark themes
- WAV export functionality confirmed working

## Files Changed
- Frontend components (App.css, Sidebar, TranscriptsView)
- Audio recorder optimization (recorder.rs)
- New audio monitoring modules
- Documentation updates